update edit product

+++ laravel-app/database/migrations/2023_03_24_125245_create_brands_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('brands', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->unsignedInteger('id');
             $table->string('name');
             $table->string('slug')->unique();
             $table->string('author', 255);
